# Sync fast path for WriteWithLength